Finished up step 47
So far in this chapter there are a few notable points:

- I moved functions like getWindowSize into the editor config.
Everywhere else is supposed to git that info from the config, so it made
sense for the function to go there.
- I didn't make the arrow constants an enum. I tried doing this, but
swift enums don't work quite the same as C enums, you can't use them
interchangably in places where we're using ints or cchars or whatever.
I'm not sure what the best solution here is, but we'll see. I see the
next steps involve messing with those bits too, so it's fine.
- There is a bug you'll notice if you move the cursor around where "~"
characters get drawn where they shouldn't. Not sure how to fix that
yet.
- I also skipped some of the steps that didn't really apply to Swift. No
need to make an expanding buffer when we can just use a String and
append to it.