Now you can save new files.
Previously, saving only worked if the editor had a filename from
something that was already pre-existing. Now if you open the editor with
no file, you'll get the chance to specify a filename and save it anyway.
Hooray! There are some bugs with how the promprt is drawn, not sure
what's causing it yet, but it *is* working.